requirements.txt
attrs==16.0.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.4.1
boto==2.40.0
bs4==0.0.1
bz2file==0.98
cffi==1.6.0
cloudpickle==0.2.1
cryptography==1.3.2
cssselect==0.9.1
decorator==4.0.10
docker-py==1.7.1
gensim==0.12.4
hiredis==0.2.0
idna==2.1
ipython==4.2.0
ipython-genutils==0.1.0
lda==1.0.3
lxml==3.6.0
mongo==0.2.0
netifaces==0.10.4
nltk==3.2.1
numpy==1.11.0
parsel==1.0.2
pbr==1.10.0
pexpect==4.1.0
pickleshare==0.7.2
ptyprocess==0.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
PyDispatcher==2.0.5
pymongo==3.2.2
pyOpenSSL==16.0.0
queuelib==1.4.2
redis==2.10.5
requests==2.10.0
retrying==1.3.3
scipy==0.17.1
Scrapy==1.1.0
service-identity==16.0.0
simplegeneric==0.8.1
six==1.10.0
smart-open==1.3.3
stop-words==2015.2.23.1
tornado==4.3
traitlets==4.2.1
Twisted==16.2.0
w3lib==1.14.2
websocket-client==0.37.0
zope.interface==4.1.3
